Isolate activateUser controller tests
